Playground/requirements.txt
python-dotenv==1.0.0
cryptography==42.0.5
fastapi==0.110.0
requests==2.31.0
pydantic==2.6.4
uvicorn==0.27.1 